UCM/MALLOC: Fix memory leak detection with valgrind
We saved pointers allocated in mmap'ed regions in a hash table in order
to release them correctly, however this makes Valgrind think the blocks
are reachable, and not fail if they leak.

The fix is to save only the page address, which breaks the link to the
allocated block, and also reduces hash size.