UCT/ROCM: send-to-self fixes
1. Cleanup handling of remote_addr vs. iov->buffer in
rocm_copy_ep_zcopy.
For put operations, src has to be iov->buffer and dst has
to be remote_addr, for get operations exactly the other way around.

2. The rkey does not necessarily contain the registration information
of the host buffer, it might be the rkey of the device buffer. Hence,
we need to check for that and locally 'lock' the host buffer using
hsa_amd_memory_lock() if the provided rkey belongs to a device buffer.

3. Introduce the ability to perform host-to-host memcpy in the
component, since the rocm/copy component can be selected for
host-to-host operations in send-to-self scenario.