UCT/ROCM: avoid repetitive calls to get_link_type
the current implementation of get_bw() in rocm/ipc uses a single value
for the bw dependent on the connection type. There is no point
in repeating the same sequence of functions calls over and over again,
the result is always the same. So use a static variable to cache the
value instead after it has been determined once.